refactor: extract shared server logic to eliminate duplication between index.ts and index-http.ts
